requirements.txt
numpy
opencv-contrib-python
Pillow
tqdm
matplotlib
selectivesearch
